+++ CHANGELOG.md
@@ -15,6 +15,7 @@
    * ADDED: documentation page for the pedestrian areas feature [#6266](https://github.com/valhalla/valhalla/pull/6266)
    * ADDED: Add use_distance to truck, bus, taxi, motorcycle and motor_scooter cost [#6214](https://github.com/valhalla/valhalla/pull/6214)
    * ADDED: limitations, future work and performance sections to the pedestrian areas documentation [#6279](https://github.com/valhalla/valhalla/pull/6279)
+   * ADDED: `location_indices` to route errors 171 and 442, the indices of the request locations that failed, in the JSON error body and on pyvalhalla's `ValhallaError` [#6281](https://github.com/valhalla/valhalla/issues/6281)
 
 ## Release Date: 2026-07-24 Valhalla 3.8.3
 * **Removed**